+++ StabilityMatrix.Avalonia/Views/Inference/InferenceImageToVideoView.axaml.cs
@@ -1,0 +1,13 @@
+﻿using StabilityMatrix.Avalonia.Controls.Dock;
+using StabilityMatrix.Core.Attributes;
+
+namespace StabilityMatrix.Avalonia.Views.Inference;
+
+[Transient]
+public partial class InferenceImageToVideoView : DockUserControlBase
+{
+    public InferenceImageToVideoView()
+    {
+        InitializeComponent();
+    }
+}