+++ CHANGELOG.md
@@ -5,11 +5,27 @@ All notable changes to Stability Matrix will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2.0.0.html).
 
-## v2.7.7
+## v2.8.0-dev.3
 ### Added
 - Added ComfyUI launch argument configs: Cross Attention Method, Force Floating Point Precision, VAE Precision
 ### Changed
 - Python Packages install dialog now allows entering multiple arguments or option flags
+
+## v2.8.0-dev.2
+### Added
+#### Inference
+- Added Image to Video project type
+#### Output Browser
+- Added support for webp files
+- Added "Send to Image to Image" and "Send to Image to Video" options to the context menu
+### Changed
+- Changed how settings file is written to disk to reduce potential data loss risk
+
+## v2.8.0-dev.1
+### Added
+#### Inference
+- Added image and model details in model selection boxes
+- Added CLIP Skip setting, toggleable from the model settings button
 
 ## v2.7.6
 ### Added